test(core): separate registry integration coverage
